ChartEntry interface
