+++ src/main/java/chart/ChartEntry.java
@@ -2,28 +2,20 @@ package chart;
 
 import java.util.Optional;
 
-import org.immutables.value.Value;
+public interface ChartEntry {
+    Integer position();
 
-@Value.Immutable
-public abstract class ChartEntry {
-    public abstract Integer position();
-    public abstract Optional<Integer> lastPosition();
-    public abstract Integer weeksOnChart();
-    public abstract String title();
-    public abstract String artist();
+    Optional<Integer> lastPosition();
 
-    @Value.Default
-    public String id() {
-        return "";
-    }
+    Integer weeksOnChart();
 
-    @Value.Default
-    public String href() {
-        return "";
-    }
+    String title();
 
-    @Value.Default
-    public String uri() {
-        return "";
-    }
+    String artist();
+
+    String id();
+
+    String href();
+
+    String uri();
 }
