+++ src/main/java/chart/CsvChartEntry.java
@@ -1,0 +1,25 @@
+package chart;
+
+import org.immutables.value.Value;
+
+@Value.Immutable
+public abstract class CsvChartEntry implements ChartEntry {
+
+    @Override
+    @Value.Default
+    public String id() {
+        return "";
+    }
+
+    @Override
+    @Value.Default
+    public String href() {
+        return "";
+    }
+
+    @Override
+    @Value.Default
+    public String uri() {
+        return "";
+    }
+}
